Corrected complex type in non-cached overlaps

+++ amset/interpolation/wavefunction.py
@@ -130,14 +130,14 @@ class WavefunctionOverlapCalculator(PeriodicLinearInterpolator):
 
 @numba.njit
 def _get_overlap(grid, data, points, n_coeffs, g_diff, g_maps):
-    initial = np.zeros(n_coeffs, dtype=np.complex64)
+    initial = np.zeros(n_coeffs, dtype=np.complex128)
     initial.real[:] = eval_linear(grid, data.real, points[0], xto.LINEAR)
     initial.imag[:] = eval_linear(grid, data.imag, points[0], xto.LINEAR)
     initial /= np.linalg.norm(initial)
     initial[:] = np.conj(initial)
 
     res = np.zeros(points.shape[0] - 1)
-    final = np.zeros(n_coeffs + 1, dtype=np.complex64)
+    final = np.zeros(n_coeffs + 1, dtype=np.complex128)
     for i in range(1, points.shape[0]):
         final.real[:-1] = eval_linear(grid, data.real, points[i], xto.LINEAR)
         final.imag[:-1] = eval_linear(grid, data.imag, points[i], xto.LINEAR)
@@ -150,7 +150,7 @@ def _get_overlap(grid, data, points, n_coeffs, g_diff, g_maps):
 
 @numba.njit
 def _get_overlap_ncl(grid, data, points, n_coeffs, g_diff, g_maps):
-    initial = np.zeros((n_coeffs, 2), dtype=np.complex64)
+    initial = np.zeros((n_coeffs, 2), dtype=np.complex128)
     initial.real[:] = eval_linear(grid, data.real, points[0], xto.LINEAR).reshape(
         (n_coeffs, 2)
     )
@@ -161,7 +161,7 @@ def _get_overlap_ncl(grid, data, points, n_coeffs, g_diff, g_maps):
     initial[:] = np.conj(initial)
 
     res = np.zeros(points.shape[0] - 1)
-    final = np.zeros((n_coeffs + 1, 2), dtype=np.complex64)
+    final = np.zeros((n_coeffs + 1, 2), dtype=np.complex128)
     for i in range(1, points.shape[0]):
         final.real[:-1] = eval_linear(grid, data.real, points[i], xto.LINEAR).reshape(
             (n_coeffs, 2)
